requirements.txt
﻿Django==5.0.14
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1

django-environ==0.11.2
django-cors-headers==4.3.1
django-filter==24.3

drf-spectacular==0.27.2

psycopg2-binary==2.9.9

gunicorn==21.2.0
Pillow==10.2.0
redis==5.0.1

asgiref==3.8.1
sqlparse==0.5.0
tzdata==2024.1
